miniS2S/requirements.txt
torch
torchaudio
pytorch-lightning
numpy
matplotlib
librosa
soundfile
h5py
pyyaml
tensorboard
tqdm
scikit-learn
Pillow
